Update fusion_blankets.py
added a bit which graphs the absorptions against thickness and 'lift density' against thickness. 

other processes can be graphed against thickness and the metric for optimisation can (and should) be changed. 

number of neutrons has been edited from 300 (I think) to 2500 to reduce statistical error on the graphs. 

to make the code run quicker (it takes around 5 mins at the moment), reduce number of neutrons and number of 'steps' in the graphing section at the end of 'main' (the final part of 'main').